created staff model and functions
